move modeldb-test-server into modeldb package
