+++ package.json
@@ -20,6 +20,7 @@
 	},
 	"workspaces": [
 		"packages/*",
+		"packages/modeldb/test/server",
 		"examples/*"
 	],
 	"devDependencies": {
