+++ packages/modeldb/test/server/tsconfig.json
@@ -21,5 +21,5 @@
 		"noFallthroughCasesInSwitch": true
 	},
 	"include": ["src"],
-	"references": [{ "path": "../modeldb-sqlite-wasm" }, { "path": "../modeldb" }]
+	"references": [{ "path": "../../../modeldb-sqlite-wasm" }, { "path": "../" }]
 }
